+++ WinQuake/gl_screen.cpp
@@ -314,9 +314,9 @@ static void SCR_CalcRefdef (void)
 	}
 
 	r_refdef.vrect.height = vid.height * size;
-	if (r_refdef.vrect.height > vid.height - sb_lines)
+	if ( static_cast<unsigned int>( r_refdef.vrect.height ) > vid.height - sb_lines)
 		r_refdef.vrect.height = vid.height - sb_lines;
-	if (r_refdef.vrect.height > vid.height)
+	if ( static_cast<unsigned int>( r_refdef.vrect.height ) > vid.height)
 			r_refdef.vrect.height = vid.height;
 	r_refdef.vrect.x = (vid.width - r_refdef.vrect.width)/2;
 	if (full)
